More interoperability to the original llama.cpp, and arguments now work
